+++ tests/document_controller/test_action_add_card.py
@@ -21,9 +21,11 @@ from hamcrest import *
 
 from mtg_proxy_printer.model.carddb import Card, MTGSet, CheckCard
 from mtg_proxy_printer.model.document_page import PageType
+from mtg_proxy_printer.model.imagedb import ImageDatabase
 from mtg_proxy_printer.document_controller import IllegalStateError
 from mtg_proxy_printer.document_controller.card_actions import ActionAddCard
 from mtg_proxy_printer.document_controller.page_actions import ActionNewPage
+from mtg_proxy_printer.units_and_sizes import CardSizes
 
 from .test_action_new_page import append_new_pages
 from .helpers import insert_card_in_page, card_container_with
@@ -31,17 +33,17 @@ from .helpers import insert_card_in_page, card_container_with
 
 @pytest.fixture()
 def card():
-    return Card("", MTGSet("", ""), "", "", "", True, "", "", True, False, 0, False, None)
+    return Card("", MTGSet("", ""), "", "", "", True, "", "", True, CardSizes.REGULAR, 0, False, None)
 
 
 @pytest.fixture()
 def oversized_card():
-    return Card("", MTGSet("", ""), "", "", "", True, "", "", True, True, 0, False, None)
+    return Card("", MTGSet("", ""), "", "", "", True, "", "", True, CardSizes.OVERSIZED, 0, False, None)
 
 
 @pytest.fixture()
-def check_card(card, image_db):
-    card.image_file = image_db.blank_image
+def check_card(card, image_db: ImageDatabase):
+    card.image_file = image_db.get_blank()
     card.is_dfc = True
     other = copy.copy(card)
     other.is_front = False